refactor: merge keyword argument checkers into a single checker
Wave 3 of the single-visitor migration. Four AST visitors that each walked
every file only to look at `[Arguments]` are collapsed into a single
`ArgumentsChecker` in the new `keyword_arguments.py` module:

- `spacing.ArgumentsChecker` (`first-argument-in-new-line`)
- `lengths.TooManyArgumentsInLineChecker` (`arguments-per-line`)
- `misc.UndefinedArgumentDefaultChecker` (`undefined-argument-default`)
- `keywords.NoEmbeddedKeywordArgumentsChecker` (`no-embedded-keyword-arguments`)

`duplicated-argument-name` also moves there from `DuplicationsChecker`. Its
`visit_Arguments` was stateless and unrelated to the duplicate tracking the
rest of that checker does, so it does not need to stay.

`undefined-argument-value` is triggered by keyword calls rather than by
`[Arguments]`, so it moves into the existing `KeywordCallChecker` instead.
This splits `UndefinedArgumentDefaultChecker` along the lines of what
actually triggers each of its two rules.

As in the previous waves the check logic moves into `check()` methods on the
rule classes, leaving the visitor responsible only for traversal.

Checkers are internal, so there is no user visible change. Verified
byte-identical output against the previous implementation over the whole
test corpus - both for the six affected rules (together and each in
isolation) and for a full `--select ALL` run covering ~189k reported issues.
